+++ sources/serveur.h
@@ -67,25 +67,25 @@ int sendHeaderField(int size, int type);
 /* isGetRequest.
  *
 */
-int isGetRequest(char* data, size_t size);
+int isGetRequest(char* data, int size);
 
 
 /* isPutRequest.
  *
 */
-int isPutRequest(char* data, size_t size);
+int isPutRequest(char* data, int size);
 
 
 /* isConnectRequest.
  *
 */
-int isConnectRequest(char* data, size_t size);
+int isConnectRequest(char* data, int size);
 
 
 /* isDeleteRequest.
  *
 */
-int isDeleteRequest(char* data, size_t size);
+int isDeleteRequest(char* data, int size);
 
 
 /*
